import front integrated partially

+++ library/src/main/java/com/kaizen/library/service/ClientService.java
@@ -5,6 +5,7 @@ import com.kaizen.library.model.Book;
 import com.kaizen.library.model.Client;
 import com.kaizen.library.repository.BookRepository;
 import com.kaizen.library.repository.ClientRepository;
+import com.kaizen.library.repository.LoanRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,9 +15,11 @@ import java.util.List;
 @Service
 public class ClientService {
 
+    private final LoanRepository loanRepository;
     private final ClientRepository clientRepository;
 
-    public ClientService(ClientRepository clientRepository) {
+    public ClientService(LoanRepository loanRepository, ClientRepository clientRepository) {
+        this.loanRepository = loanRepository;
         this.clientRepository = clientRepository;
     }
 
@@ -29,12 +32,16 @@ public class ClientService {
     }
 
     public void delete(Long id) {
+       boolean hasLoan = loanRepository.findAll().stream().anyMatch(loan -> loan.getClient().getId().equals(id));
+        if (hasLoan) {
+            throw new IllegalStateException("Client has active loans and cannot be deleted.");
+        }
         clientRepository.deleteById(id);
     }
 
 
     public Client findById(Long clientId) {
         return clientRepository.findById(clientId)
-                .orElseThrow(() -> new ClientNotFoundException());
+                .orElseThrow(ClientNotFoundException::new);
     }
 }
